+++ src/ImageSharp/Processing/Processors/PixelShading/DelegatePixelShaderProcessor.cs
@@ -5,7 +5,7 @@ using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing.Delegates;
 using SixLabors.Primitives;
 
-namespace SixLabors.ImageSharp.Processing.Processors
+namespace SixLabors.ImageSharp.Processing.Processors.PixelShading
 {
     /// <summary>
     /// Applies a user defined pixel shader effect through a given delegate.
