+++ src/ImageSharp/Processing/Processors/PixelShading/DelegatePixelShaderProcessor{TPixel}.cs
@@ -10,7 +10,7 @@ using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing.Delegates;
 using SixLabors.Primitives;
 
-namespace SixLabors.ImageSharp.Processing.Processors
+namespace SixLabors.ImageSharp.Processing.Processors.PixelShading
 {
     /// <summary>
     /// Performs simple binary threshold filtering against an image.
